Rename package setting -> game
