+++ ladder/jaeeun-na-1/src/main/java/org/ooptraining/game/GameContext.java
@@ -1,4 +1,4 @@
-package org.ooptraining.setting;
+package org.ooptraining.game;
 
 import lombok.Getter;
 import lombok.ToString;
