+++ ladder/jaeeun-na-1/src/main/java/org/ooptraining/game/Parser.java
@@ -1,4 +1,4 @@
-package org.ooptraining.setting;
+package org.ooptraining.game;
 
 public interface Parser {
     void parse(final String line, final GameContext.Builder builder);
