+++ ladder/jaeeun-na-1/src/main/java/org/ooptraining/render/policy/RandomRenderPolicy.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.ooptraining.Participant;
 import org.ooptraining.render.RenderContext;
 import org.ooptraining.render.RenderPolicy;
-import org.ooptraining.setting.GameContext;
+import org.ooptraining.game.GameContext;
 
 import java.util.List;
