+++ ladder/jaeeun-na-1/src/test/java/org/ooptraining/render/RendererTest.java
@@ -3,7 +3,7 @@ package org.ooptraining.render;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.ooptraining.Participant;
-import org.ooptraining.setting.GameContext;
+import org.ooptraining.game.GameContext;
 
 import java.util.Arrays;
 import java.util.List;
