+++ ladder/jaeeun-na-1/src/test/java/org/ooptraining/util/dummy/data/SettingContextDummys.java
@@ -2,9 +2,8 @@ package org.ooptraining.util.dummy.data;
 
 import org.junit.jupiter.params.provider.Arguments;
 import org.ooptraining.Participant;
-import org.ooptraining.setting.GameContext;
+import org.ooptraining.game.GameContext;
 
-import java.util.Arrays;
 import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
